ref(core-app): split oversized plugin modules
